+++ packages/shared/src/types.ts
@@ -107,6 +107,7 @@ export interface BaseAuth {
     variables?: I18nVariables
   }
   theme?: 'default' | string
+  passwordLimit?: boolean
 }
 
 export type I18nVariables = {