+++ packages/solid/src/components/Auth/interfaces/UpdatePassword.tsx
@@ -7,6 +7,7 @@ import { Button, Container, Input, Label, Message } from '../../UI'
 function UpdatePassword(props: {
   supabaseClient: SupabaseClient
   i18n: I18nVariables
+  passwordLimit: boolean
   appearance?: Appearance
 }) {
   const [password, setPassword] = createSignal('')
@@ -19,6 +20,11 @@ function UpdatePassword(props: {
     setError('')
     setMessage('')
     setLoading(true)
+    if (props.passwordLimit && password().length > 72) {
+      setError('Password exceeds maxmium length of 72 characters')
+      setLoading(false)
+      return
+    }
     const { error } = await props.supabaseClient.auth.updateUser({
       password: password(),
     })
@@ -47,7 +53,9 @@ function UpdatePassword(props: {
               id="password"
               name="password"
               autofocus
-              placeholder={props.i18n?.update_password?.password_label}
+              placeholder={
+                props.i18n?.update_password?.password_input_placeholder
+              }
               type="password"
               onkeyup={(e) => setPassword(e.currentTarget.value)}
               appearance={props.appearance}